requirements.txt
anthropic
mss
requests
Pillow
